Use RootLinks instead of confusing Gdc
