+++ src/main/java/com/gooddata/gdc/DataStoreService.java
@@ -69,7 +69,7 @@ public class DataStoreService {
 
     private UriPrefixer getPrefixer() {
         if (prefixer == null) {
-            final String uriString = gdcService.getGdc().getUserStagingLink();
+            final String uriString = gdcService.getRootLinks().getUserStagingLink();
             final URI uri = URI.create(uriString);
             prefixer = new UriPrefixer(uri.isAbsolute() ? uri : gdcUri.resolve(uriString));
             sardine.enablePreemptiveAuthentication(prefixer.getUriPrefix().getHost());
