+++ src/main/java/com/gooddata/gdc/GdcService.java
@@ -21,12 +21,27 @@ public class GdcService extends AbstractService {
      * Obtains GoodData API root links.
      *
      * @return GoodData API root links
+     * @deprecated use {@link #getRootLinks()} instead
      */
+    @Deprecated
     public Gdc getGdc() {
         try {
             return restTemplate.getForObject(Gdc.URI, Gdc.class);
         } catch (GoodDataException | RestClientException e) {
             throw new GoodDataException("Unable to get gdc about", e);
+        }
+    }
+
+    /**
+     * Obtains GoodData API root links.
+     *
+     * @return GoodData API root links
+     */
+    public RootLinks getRootLinks() {
+        try {
+            return restTemplate.getForObject(RootLinks.URI, RootLinks.class);
+        } catch (GoodDataException | RestClientException e) {
+            throw new GoodDataException("Unable to get gdc root links", e);
         }
     }
